Pull out generic component into a base class for wx implementations
